server/requirements.txt
# Agent Assist Console - Backend Requirements

# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0

# Google GenAI
google-genai==1.55.0

# Data Processing
pandas==2.2.3
openpyxl==3.1.5
fuzzywuzzy==0.18.0
python-Levenshtein==0.26.0

# HTTP Client
aiohttp==3.11.0

# Environment Management
python-dotenv==1.0.1

# Optional: Development tools
# pytest==8.3.0
# pytest-asyncio==0.24.0
# black==24.10.0
# flake8==7.1.0
    